client side dnd item list flow almost complete
