+++ components/composition/deck-list.tsx
@@ -1,97 +1,145 @@
 "use client";
 
-import DNDSortableList, {
-  BaseItem,
-  DisplayAttributes,
-} from "../primitive/drag-drop-list";
-import {
-  DndContext,
-  DragEndEvent,
-  DragOverEvent,
-  DragStartEvent,
-  UniqueIdentifier,
-} from "@dnd-kit/core";
-import { SortableData } from "@dnd-kit/sortable";
-import { useId, useState } from "react";
+import { BaseItem } from "../primitive/drag-drop-list";
+import ItemList from "./item-list";
+import { useState } from "react";
 
-export default function DeckList<T extends BaseItem>(props: {
-  initialItems: BaseItem[];
-  attributes: DisplayAttributes<T>;
-}) {
-  const id = useId();
-  const [items, setItems] = useState(props.initialItems);
-  const [wasOutside, setWasOutside] = useState(false);
-  // const changedItems = [];
+export default function DeckList() {
+  const [decks, setDecks] = useState<BaseItem[]>(recalculateDecks(tempDecks));
 
-  function swapGroup(group: BaseItem[], activeIdx: number, overIdx: number) {
-    const item = group.splice(activeIdx, 1)[0];
-    group.splice(overIdx, 0, item);
+  function recalculateDecks(items: BaseItem[]) {
+    items.forEach((item) => {
+      if (item.items && item.items.length > 0) {
+        item.new = 0;
+        item.learning = 0;
+        item.review = 0;
+        item.items.forEach((innerItem) => {
+          item.new += innerItem.new;
+          item.learning += innerItem.learning;
+          item.review += innerItem.review;
+        });
+      }
+    });
+
+    return items;
   }
 
-  function sameGroupSwap(
-    parentId: UniqueIdentifier,
-    activeIdx: number,
-    overIdx: number,
-  ) {
-    if (typeof parentId !== "string") return;
-    const parentItemId = parentId.split(".")[1];
-
-    const group =
-      parentItemId === "root"
-        ? items
-        : items.find((item) => item.id === parentItemId)?.items;
-
-    if (!group) return;
-
-    swapGroup(group, activeIdx, overIdx);
-
-    setItems([...items]);
-  }
-
-  function handleDragStart(e: DragStartEvent) {
-    console.log(e);
-    //  todo
-  }
-
-  function handleDragEnd(e: DragEndEvent) {
-    const activeSortable = e.active.data.current as SortableData;
-    const overSortable = e.over?.data.current as SortableData;
-
-    if (!activeSortable || !overSortable) {
-      return;
-    }
-    console.log(wasOutside);
-    if (!wasOutside) {
-      sameGroupSwap(
-        overSortable.sortable.containerId,
-        activeSortable.sortable.index,
-        overSortable.sortable.index,
-      );
-    }
-    setWasOutside(false);
-    // setItems([...items]);
-  }
-
-  function handleDragOver(e: DragOverEvent) {
-    // const overId = e.over?.id;
-    // const isOutside = items.some((item) => item.id === overId);
-    //
-    // if (isOutside && !wasOutside && items[0].items!.length > 0) {
-    //   // const item = items[0].items!.shift();
-    //   // items.push(item!);
-    // }
-    //
-    // setWasOutside(isOutside);
+  function recalculateAndSetDecks(items: BaseItem[]) {
+    setDecks([...recalculateDecks(items)]);
   }
 
   return (
-    <DndContext
-      id={id}
-      onDragStart={handleDragStart}
-      onDragEnd={handleDragEnd}
-      onDragOver={handleDragOver}
-    >
-      <DNDSortableList items={items} attributes={props.attributes} />
-    </DndContext>
+    <ItemList
+      items={decks}
+      setItems={recalculateAndSetDecks}
+      attributes={{
+        Title: "title",
+        New: "new",
+        Learning: "learning",
+        Review: "review",
+      }}
+    />
   );
 }
+
+const tempDecks = [
+  {
+    id: "group-0",
+    title: "First Group",
+    new: 0,
+    learning: 0,
+    review: 0,
+    index: 8,
+    isGroup: true,
+    items: [
+      {
+        id: "f3b81aaa-6033-42f9-a340-0979e5410f80",
+        title: "Crocuta crocuta",
+        new: 35,
+        learning: 32,
+        review: 347,
+        index: 0,
+      },
+      {
+        id: "c149aaaa-07a5-4bd7-99e2-733ad3bc8008",
+        title: "Phascogale calura",
+        new: 81,
+        learning: 183,
+        review: 309,
+        index: 5,
+      },
+      {
+        id: "fb3aaa7f-3604-4d32-a503-f1ba1e90b8a7",
+        title: "Anas punctata",
+        new: 5,
+        learning: 11,
+        review: 182,
+        index: 2,
+      },
+    ],
+  },
+  {
+    id: "f3b8180b-6033-42f9-a340-0979e5410f80",
+    title: "Crocuta crocuta",
+    new: 35,
+    learning: 32,
+    review: 347,
+    index: 0,
+  },
+  {
+    id: "c149a4a6-07a5-4bd7-99e2-733ad3bc8008",
+    title: "Phascogale calura",
+    new: 81,
+    learning: 183,
+    review: 309,
+    index: 5,
+  },
+  {
+    id: "fb39c57f-3604-4d32-a503-f1ba1e90b8a7",
+    title: "Anas punctata",
+    new: 5,
+    learning: 11,
+    review: 182,
+    index: 2,
+  },
+  {
+    id: "6507b961-1cfc-4dd1-a487-7171954df091",
+    title: "Marmota flaviventris",
+    new: 83,
+    learning: 99,
+    review: 134,
+    index: 3,
+  },
+  {
+    id: "6b509de1-14ad-4e91-889e-1f878f2ddaaa",
+    title: "Chlidonias leucopterus",
+    new: 45,
+    learning: 122,
+    review: 41,
+    index: 4,
+  },
+  {
+    id: "11e9db11-e946-4ea5-91f2-86663bc46a67",
+    title: "Otaria flavescens",
+    new: 94,
+    learning: 88,
+    review: 415,
+    index: 1,
+  },
+  {
+    id: "ce052d2b-9ff1-4ed3-b3e0-3008247fcc16",
+    title: "Cebus apella",
+    new: 24,
+    learning: 179,
+    review: 368,
+    index: 6,
+  },
+  {
+    id: "99bb4d5a-90f9-4f8a-8c2e-b9df30792d8d",
+    title: "Passer domesticus",
+    new: 50,
+    learning: 42,
+    review: 173,
+    index: 7,
+  },
+];
